# docs/product/case-workspace/api/openapi.yaml
openapi: 3.0.3
...
info:
  title: Consultify — Case Workspace API
  version: 1.0.0-as-built-2026-08-10
  description: |
    HTTP contract for the 11 Case Workspace domain services.

    **Envelope.** Every success response is `{ "data": <payload> }`. The one
    exception is `GET /process-definitions`, which is the only natively
    paginated list in the codebase and returns
    `{ "data": [...], "nextCursor": string|null }`.

    **Errors.** Every error response is produced by
    `server/src/utils/ErrorHandler.ts#errorHandlerMiddleware` and is shaped
    `{ "status": "fail"|"error", "correlationId": string|null,
       "error": { "code": string, "message": string, ... } }`.
    `error.code` is the SCREAMING_SNAKE upper-casing of the domain service's
    own `snake_case` error code (`_shared/errors.ts#toCaseWorkspaceAppError`),
    so it is stable across releases and safe to branch on.

    **Correlation.** An inbound `X-Correlation-ID` request header is honoured
    and echoed; when absent, one is minted per request
    (`_shared/errors.ts#resolveCorrelationId`).

    **Optimistic concurrency.** Mutations on versioned aggregates take
    `expectedVersion` in the request BODY (not an `If-Match` header). A stale
    value yields `409` with a `*_VERSION_CONFLICT` code and no mutation.

    **Idempotency.** Where a command has an idempotency primitive, the route
    reads the `Idempotency-Key` request header, falling back to a body field.
...
servers:
  - url: /api/v8/case-workspace
...
security:
  - bearerAuth: []
...
paths:
...
  /case-intake/work-orders/propose:
    post:
      tags: [Case intake]
      operationId: proposeWorkOrder
      summary: Digest a work order for human confirmation. Creates NOTHING.
      description: |
        Returns `200`, not `201` — CW-CANON-01: proposing creates no Case and no
        Run, and `data.caseCreated`/`data.runCreated` are always `false` as an
        assertable statement of that. `organizationId` is forced to the
        authenticated org: it is part of the digested payload, so a
        client-supplied value would let a caller mint a digest for a tenant
        they do not belong to.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workOrder]
              properties:
                workOrder: { $ref: '#/components/schemas/WorkOrderDraft' }
      responses:
        '200':
          description: Work order digested. Nothing was created.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/ProposeWorkOrderResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/OrgRoleDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /case-intake/work-orders/confirm:
    post:
      tags: [Case intake]
      operationId: confirmWorkOrder
      summary: Confirm a previously proposed work order; creates exactly one Case.
      description: |
        The status code IS the answer: `201` when this call created the Case,
        `200` when an existing Case was reused (idempotent retry, or a Case
        already existed for the project) — so a client distinguishes the two
        without parsing the body. `confirmedDigest` must be the verbatim
        `sha256:<64 hex>` returned by the propose step; anything else fails
        `422 INTAKE_WORK_ORDER_DIGEST_MISMATCH`. Confirming never starts work
        (`data.runCreated` is always `false`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workOrder, confirmedDigest]
              properties:
                workOrder: { $ref: '#/components/schemas/WorkOrderDraft' }
                confirmedDigest:
                  type: string
                  pattern: '^sha256:[0-9a-f]{64}$'
      responses:
        '200':
          description: An existing Case was reused.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/ConfirmWorkOrderResult' }
        '201':
          description: Exactly one Case was created by this call.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/ConfirmWorkOrderResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/OrgRoleDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /case-intake/cases/{caseId}/work-orders:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [Case intake]
      operationId: getConfirmedWorkOrders
      summary: The confirmed work orders behind one Case.
      description: |
        Access is gated inside the service by `requireCaseAccess`, so an unknown
        and a cross-tenant caseId both answer `404`.
      responses:
        '200':
          description: Confirmed work order records.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ConfirmedWorkOrderRecord' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ===========================================================================
  # cases.routes.ts — caseCoreService
  # ===========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Verified by `verifyToken` and `requireV8OrgContext` upstream of this
        router (routes/v8/index.ts:55-56, 83). The case-workspace handlers
        themselves never parse a token — they read the already-populated
        `req.v8Context`.
  parameters:
    CaseId:
      { in: path, name: caseId, required: true, schema: { type: string, minLength: 1 } }
...
  responses:
...
    ValidationError:
      description: |
        Request body / query / path parameters failed zod validation, or a
        required idempotency key was absent. `error.code` is
        `VALIDATION_ERROR`; `error.issues[]` carries `{path,message,code}` per
        failure (see `_shared/validate.ts`). Some routes emit a bare
        `{ error: { code, message } }` for the missing-idempotency-key case
        without going through the error middleware.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    OrgRoleDenied:
      description: |
        `not_org_member` or `insufficient_org_role` (403), or a genuine
        caller-identity authorization failure (`SELF_APPROVAL_FORBIDDEN`,
        `PLAN_DIFF_CROSS_CASE_FORBIDDEN`,
        `PROCESS_DEFINITION_SHARE_NOT_AUTHORIZED`). NOT used for cross-tenant
        resource access — that is 404.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: |
        The resource does not exist, OR it exists in a different tenant
        (`case_access_denied` → 404, SEC-009 enumeration safety).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Unprocessable:
      description: |
        Cross-reference / cross-tenant consistency failure between
        caller-supplied ids (`*_MISMATCH`, `*_INVALID`), or a business rule
        this codebase classifies as unprocessable. Also the fallback status for
        any unrecognized `snake_case` domain error code.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    InternalError:
      description: |
        Unexpected failure, or a genuine internal invariant breach
        (`*_INSERT_FAILED`, `*_IDEMPOTENCY_RECORD_NOT_FOUND`). Never leaks a
        stack trace or SQL text in production.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        status: { type: string, enum: [fail, error] }
        correlationId: { type: string, nullable: true }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                SCREAMING_SNAKE upper-casing of the domain service's own
                snake_case code, or `VALIDATION_ERROR` / `INTERNAL_ERROR`.
            message: { type: string }
            issues:
              type: array
              description: Present on VALIDATION_ERROR only.
              items:
                type: object
                properties:
                  path: { type: string }
                  message: { type: string }
                  code: { type: string }
          additionalProperties: true

    # -- shared request bodies --------------------------------------------------
...
    CaseProfile: { type: string, enum: [LIGHT, STANDARD, TRANSFORMATION, MONITORING] }
    GovernanceTier: { type: string, enum: [LIGHTWEIGHT, STANDARD, CONTROLLED] }
    AutonomyPolicy: { type: string, enum: [ASK_EACH_ACTION, ASK_MATERIAL_ACTIONS, EXECUTE_APPROVED_PLAN] }
...
    ClosureType:
      type: string
      enum: [DELIVERY_COMPLETED, DECISION_COMPLETED, IMPLEMENTATION_COMPLETED, OUTCOME_VALIDATED, COMPLETED_PARTIAL]
...
    WorkOrderDraft:
      type: object
      description: |
        NOTE what is NOT trimmed by the route: `goal`, `scope[]`,
        `expectedOutcome`, `constraints[]` and `successCriteria[]` pass through
        exactly as sent. Normalization (NFC + trim) happens inside the service,
        in the same step as digesting, so the digested string is always the
        string the service saw.
      required: [projectId, goal, scope, expectedOutcome, contractedClosureType]
      properties:
        projectId: { type: string, minLength: 1 }
        goal: { type: string, minLength: 1 }
        scope:
          type: array
          minItems: 1
          items: { type: string }
          description: Ordered — the order is part of what the human read and is digested.
        expectedOutcome: { type: string, minLength: 1 }
        constraints: { type: array, items: { type: string }, nullable: true }
        successCriteria: { type: array, items: { type: string }, nullable: true }
        contractedClosureType: { $ref: '#/components/schemas/ClosureType' }
        caseProfile:
          allOf: [{ $ref: '#/components/schemas/CaseProfile' }]
          nullable: true
        governanceTier:
          allOf: [{ $ref: '#/components/schemas/GovernanceTier' }]
          nullable: true
        autonomyPolicy:
          allOf: [{ $ref: '#/components/schemas/AutonomyPolicy' }]
          nullable: true
        sourceConversationId: { type: string, minLength: 1, nullable: true }
        sourceMessageId: { type: string, minLength: 1, nullable: true }
      additionalProperties: false
    CanonicalWorkOrder:
      type: object
      description: The normalized, digested form. Optional inputs are defaulted, never null.
      required:
        [workOrderSchemaVersion, organizationId, projectId, goal, scope, expectedOutcome,
         constraints, successCriteria, contractedClosureType, caseProfile, governanceTier,
         autonomyPolicy]
      properties:
        workOrderSchemaVersion: { type: integer }
        organizationId: { type: string }
        projectId: { type: string }
        goal: { type: string }
        scope: { type: array, items: { type: string } }
        expectedOutcome: { type: string }
        constraints: { type: array, items: { type: string } }
        successCriteria: { type: array, items: { type: string } }
        contractedClosureType: { $ref: '#/components/schemas/ClosureType' }
        caseProfile: { $ref: '#/components/schemas/CaseProfile' }
        governanceTier: { $ref: '#/components/schemas/GovernanceTier' }
        autonomyPolicy: { $ref: '#/components/schemas/AutonomyPolicy' }
        sourceConversationId: { type: string, nullable: true }
        sourceMessageId: { type: string, nullable: true }
    ProposeWorkOrderResult:
      type: object
      required:
        [workOrderId, workOrderDigest, workOrder, proposedEventId, alreadyProposed,
         caseCreated, runCreated]
      properties:
        workOrderId:
          type: string
          description: Deterministic `wo-<sha256 hex>` — same content, same id, forever.
        workOrderDigest:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
          description: What the human must confirm, verbatim.
        workOrder: { $ref: '#/components/schemas/CanonicalWorkOrder' }
        proposedEventId: { type: string }
        alreadyProposed: { type: boolean }
        caseCreated: { type: boolean, enum: [false], description: Always false (CW-CANON-01). }
        runCreated: { type: boolean, enum: [false], description: Always false. }
    ConfirmWorkOrderResult:
      type: object
      required:
        [caseId, workOrderId, workOrderDigest, caseCreated, reused, reuseReason,
         confirmedEventId, confirmedEventDeduplicated, runCreated]
      properties:
        caseId: { type: string }
        workOrderId: { type: string }
        workOrderDigest: { type: string }
        caseCreated:
          type: boolean
          description: |
            `true` = THIS call created the Case (HTTP 201); `false` = an
            existing Case was reused (HTTP 200). `reused` is its inverse.
        reused: { type: boolean }
        reuseReason:
          type: string
          enum: [none, work_order_already_confirmed, existing_case_for_project]
        confirmedEventId: { type: string }
        confirmedEventDeduplicated: { type: boolean }
        runCreated: { type: boolean, enum: [false], description: Always false — confirming does not start work. }
    ConfirmedWorkOrderRecord:
      type: object
      required: [eventId, caseId, workOrderId, workOrderDigest, confirmedByActorId, confirmedAt, summary]
      properties:
        eventId: { type: string }
        caseId: { type: string }
        workOrderId: { type: string }
        workOrderDigest: { type: string }
        confirmedByActorId: { type: string }
        confirmedAt: { type: string }
        summary: { type: object, additionalProperties: true }
...